+++ SwaggerGen/Swagger/Type/StringUuidType.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace SwaggerGen\Swagger\Type;
+
+/**
+ * RFC 4122 UUID type definition
+ *
+ * @package    SwaggerGen
+ * @author     Martijn van der Lee <martijn@vanderlee.com>
+ * @copyright  2014-2016 Martijn van der Lee
+ * @license    https://opensource.org/licenses/MIT MIT
+ */
+class StringUuidType extends AbstractRegexType
+{
+
+	/**
+	 * Construct and setup the regular expression for this type
+	 * 
+	 * @param \SwaggerGen\Swagger\AbstractObject $parent
+	 * @param string $definition
+	 */
+	public function __construct(\SwaggerGen\Swagger\AbstractObject $parent, $definition)
+	{
+		parent::__construct($parent, $definition, 'uuid', '[a-f0-9]{8}-[a-f0-9]{4}-[1-5][a-f0-9]{3}-[89ab][a-f0-9]{3}-[a-f0-9]{12}');
+	}
+
+}