+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.3.8 - 2016-11-16
+### Added
+- New string-based builtin type `uuid`.
+- Allow model definitions to overwrite builtin types.
+
 ## 2.3.7 - 2016-10-23
 ### Fixed
 - Supports referencing a single `definition` (a.k.a. `model`) from within